:ambulance: Fix: Repository Test 수정
