+++ src/test/java/com/waterfogsw/voucher/voucher/VoucherMemoryRepositoryTests.java
@@ -8,41 +8,93 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import java.lang.reflect.Field;
+import java.util.List;
+import java.util.Map;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.samePropertyValuesAs;
 
 public class VoucherMemoryRepositoryTests {
 
-    private VoucherRepository voucherRepository = new VoucherMemoryRepository();
+    private VoucherRepository repository = new VoucherMemoryRepository();
 
     @Nested
     @DisplayName("save 메소드는")
     class Describe_save {
         @Nested
-        @DisplayName("인자로 전달받은 Voucher 가 null 이면")
-        class Context_with_null_argument {
+        @DisplayName("저장할 바우처의 기본키가 중복되지 않으면")
+        class Context_not_duplicate_prime_key {
 
             @Test
-            @DisplayName("IllegalArgumentException 예외를 발생시킨다")
-            void it_throw_error() {
-                assertThrows(IllegalArgumentException.class, () -> voucherRepository.save(null));
+            @DisplayName("성공한다")
+            void it_not_throw_error() {
+                final Voucher voucher1 = new FixedAmountVoucher(1L, 1000);
+                final Voucher voucher2 = new FixedAmountVoucher(2L, 2000);
+
+                final Voucher savedVoucher1 = repository.save(voucher1);
+                final Voucher savedVoucher2 = repository.save(voucher2);
+
+                assertThat(savedVoucher1, samePropertyValuesAs(voucher1));
+                assertThat(savedVoucher2, samePropertyValuesAs(voucher2));
             }
         }
 
         @Nested
-        @DisplayName("Voucher 가 정상적으로 저장되면")
-        class Context_with_saved_success {
+        @DisplayName("저장할 바우처의 기본키가 중복되면")
+        class Context_duplicate_prime_key {
 
             @Test
-            @DisplayName("저장한 Voucher 를 return 한다")
-            void it_return_saved_voucher() {
-                Voucher voucher = new FixedAmountVoucher(1000);
+            @DisplayName("데이터가 업데이트 된다")
+            void it_not_throw_error() {
+                final Voucher voucher1 = new FixedAmountVoucher(1L, 1000);
+                final Voucher voucher2 = new FixedAmountVoucher(1L, 2000);
 
-                Voucher savedVoucher = voucherRepository.save(voucher);
+                final Voucher savedVoucher1 = repository.save(voucher1);
+                final Voucher savedVoucher2 = repository.save(voucher2);
 
-                assertThat(savedVoucher.getType(), is(savedVoucher.getType()));
-                assertThat(savedVoucher.getValue(), is(savedVoucher.getValue()));
+                assertThat(savedVoucher2, samePropertyValuesAs(voucher2));
+            }
+        }
+    }
+
+    @Nested
+    @DisplayName("findAll 메서드는")
+    class Describe_findAll {
+
+        @Nested
+        @DisplayName("호출되면")
+        class Context_with_call {
+
+            @Test
+            @DisplayName("저장된 모든 바우처에 대한 리스트를 반환한다")
+            void it_return_list() {
+                Field voucherStore;
+
+                try {
+                    voucherStore = repository.getClass().getDeclaredField("voucherStore");
+                    voucherStore.setAccessible(true);
+                } catch (NoSuchFieldException testException) {
+                    throw new RuntimeException(testException.getMessage());
+                }
+
+                try {
+                    Map<Long, Voucher> map = (Map<Long, Voucher>) voucherStore.get(repository);
+
+                    final Voucher voucher1 = new FixedAmountVoucher(1L, 1000);
+                    final Voucher voucher2 = new FixedAmountVoucher(2L, 1000);
+
+                    map.put(voucher1.getId(), voucher1);
+                    map.put(voucher2.getId(), voucher2);
+
+                    final List<Voucher> voucherList = repository.findAll();
+
+                    assertThat(voucherList.size(), is(2));
+
+                } catch (IllegalAccessException testException) {
+                    throw new RuntimeException(testException.getMessage());
+                }
             }
         }
     }
